improve retry functionality
